requirements.txt
pip==23.2.1
absl-py==1.4.0
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
exceptiongroup==1.1.3
fastapi==0.101.1
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
keras==2.13.1
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
mpmath==1.3.0
numpy==1.21.6
oauthlib==3.2.2
onnxruntime==1.14.1
opencv-python==4.2.0.32
opt-einsum==3.3.0
packaging==21.3
Pillow==9.3.0
protobuf==4.21.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.0
pydantic_core==2.6.3
pyreadline3==3.4.1
python-dotenv==1.0.0
python-multipart==0.0.5
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.10.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.5.0
urllib3==1.26.16
uvicorn==0.20
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.15.0
zipp==3.16.2
